requirements.txt
Pillow
2captcha-python
selenium
Appium-Python-Client
2captcha-python
